fix: eslint frontend is done
