+++ .husky/pre-commit
@@ -3,8 +3,8 @@
 
 npx validate-branch-name
 # TODO: fix eslint errors on client and server sides
-# cd client
-# npm run lint:script
+ cd client
+ npm run lint:script
 # npm run lint:styles
 # cd ../server
 # npm run lint
