+++ client/src/components/modals/ModalQuestionnaireAnswers.vue
@@ -22,7 +22,7 @@
               disabled
               class="input-answer"
               type="text"
-              :placeholder="form_field.requisition_field[0].value"
+              :placeholder="form_field.requisition_field ? form_field.requisition_field[0].value : '-' "
             />
           </div>
         </div>
@@ -39,12 +39,14 @@
 <script setup lang="ts">
 import { onBeforeMount, ref, watch } from "vue";
 import { useFormStore } from "@/store/form_store";
+import type { Ref } from "vue";
+import type {IForm} from "@/store/models/forms/form.model";
 
 const formStore = useFormStore();
-const formAnswers = ref();
+const formAnswers:Ref<IForm> = ref({});
 
 const props = defineProps<{
-  requisition: any;
+  requisition: IForm;
 }>();
 
 watch(
@@ -60,7 +62,7 @@ onBeforeMount(async () => {
 
 async function fetchRequisitionAnswers() {
   formAnswers.value = await formStore.fetchRequisitionAnswers(
-    props.requisition.id,
+    props.requisition?.id ?? -1,
   );
 }
 </script>
