+++ client/src/views/teams/QuestionnairePage.vue
@@ -1,23 +1,23 @@
 <template>
-  <div class="form">
-    <div class="wrapper-question" v-for="(form, index) in data">
-      <div class="question-label">
-        <div class="num-question">Вопрос {{ index + 1 }}</div>
-        <label class="checkbox-label">
-          <input type="checkbox" v-model="form.required" />
-          <div class="checkbox-custom"></div>
-          <span class="checkbox-text">Обязательный</span>
-        </label>
-      </div>
-      <div class="wrap-input">
-        <textarea class="input-question" v-model="form.title" />
-        <button class="remove-btn" @click="removeInput(index)">Удалить</button>
-      </div>
-    </div>
-    <button class="add-btn" @click="addInput">Добавить вопрос</button>
+    <div class="form">
+        <div class="wrapper-question" v-for="(field, index) in fields" v-bind:key="index">
+            <div class="question-label">
+                <div class="num-question">Вопрос {{ index + 1 }}</div>
+                <label class="checkbox-label">
+                    <input type="checkbox" v-model="field.required"/>
+                    <div class="checkbox-custom"></div>
+                    <span class="checkbox-text">Обязательный</span>
+                </label>
+            </div>
+            <div class="wrap-input">
+                <textarea class="input-question" v-model="field.title"/>
+                <button class="remove-btn" @click="removeInput(index)">Удалить</button>
+            </div>
+        </div>
+        <button class="add-btn" @click="addInput">Добавить вопрос</button>
 
-    <button class="save-btn" @click="submit">Сохранить</button>
-  </div>
+        <button class="save-btn" @click="submit">Сохранить</button>
+    </div>
 </template>
 
 <script lang="ts" setup>
@@ -25,6 +25,8 @@ import axios from "axios";
 import {onBeforeMount, ref} from "vue";
 import {useFormStore} from "@/store/form_store";
 import {useRoute} from "vue-router";
+import type {IFormField} from "@/store/models/forms/form-field.model";
+import type {Ref} from "vue";
 
 const route = useRoute();
 
@@ -34,130 +36,108 @@ let idForm: number;
 
 const formStore = useFormStore();
 
-const data = ref();
+const fields:Ref<IFormField[]> = ref([]);
 const deletedFields = ref();
 
 const responseMsg = ref();
 
 onBeforeMount(async () => {
-  await fetchFormFields();
-  deletedFields.value = [];
+    await fetchFormFields();
+    deletedFields.value = [];
 });
 
 async function fetchFormFields() {
-  data.value = await formStore.fetchFormFields(idTeam);
-  idForm = await formStore.fetchFormId(idTeam);
+    fields.value = await formStore.fetchFormFields(idTeam) as IFormField[];
+    idForm = await formStore.fetchFormId(idTeam);
 }
 
 const addInput = () => {
-  if (!data.value) {
-    data.value = []; // Initialize data.value as an empty array
-  }
-  data.value.push({ title: "", required: true });
+    if (!fields.value) {
+        fields.value = []; // Initialize data.value as an empty array
+    }
+    fields.value.push({title: "", required: true});
 };
 
 const submit = async () => {
-  const promises = [];
+    const promises = [];
 
-  for (let i = data.value.length - 1; i >= 0; i--) {
-    const form = data.value[i];
-    // if (!form.title.trim()) {
-    //   removeInput(i);
-    // } else {
-    if (form.id == null) {
-      const promise = createFormFields(form.title, idForm, form.required);
-      promises.push(promise);
+    for (let i = fields.value.length - 1; i >= 0; i--) {
+        const field = fields.value[i];
+
+        if (field.id == null && field.title && field.required) {
+            const promise = createFormFields(field.title, idForm, field.required);
+            promises.push(promise);
+        }
     }
-  }
 
-  // const results = await Promise.all(promises);
-  // const idFields = results.filter((result) => result !== null).join("\n");
-
-  if (idForm == undefined) {
-    await createForm(idTeam);
-  } else {
-    await archiveFields(deletedFields.value, true);
-
-    // updateForm(idForm)
-  }
-  //console.log(idFields);
+    if (idForm == undefined) {
+        await createForm(idTeam);
+    } else {
+        await archiveFields(deletedFields.value, true);
+        // updateForm(idForm)
+    }
 };
 
 const removeInput = (index: number) => {
-  if (data.value[index]) {
-    deletedFields.value.push(data.value[index]);
-    data.value.splice(index, 1);
-  }
+    if (fields.value[index]) {
+        deletedFields.value.push(fields.value[index]);
+        fields.value.splice(index, 1);
+    }
 };
 
 async function createFormFields(
-  title: string,
-  form_id: number,
-  required: boolean,
+    title: string,
+    form_id: number,
+    required: boolean,
 ) {
-  responseMsg.value = "сохранено";
+    responseMsg.value = "сохранено";
 
-  try {
-    const response = await axios.post("/api/forms/field", {
-      title: title,
-      required: required,
-      form: form_id,
-    });
-    return response.data.id.toString();
-  } catch (error) {
-    if (error.response) {
-      responseMsg.value = error.response.data.message;
+    try {
+        const response = await axios.post("/api/forms/field", {
+            title: title,
+            required: required,
+            form: form_id,
+        });
+        return response.data.id.toString();
+    } catch (error) {
+        if (error.response) {
+            responseMsg.value = error.response.data.message;
+        }
+        return null;
     }
-    return null;
-  }
 }
 
 async function createForm(idTeam: number) {
-  responseMsg.value = "сохранено";
+    responseMsg.value = "сохранено";
 
-  await axios
-    .post("/api/forms", {
-      team_id: idTeam,
-    })
-    .catch((err) => {
-      if (err.response) {
-        responseMsg.value = err.response.data.message;
-      }
-    });
-}
-
-async function archiveFields(deletedFields: [], is_archive: boolean) {
-  console.log("deletedFields");
-  console.log(deletedFields);
-  for (let i = 0; i < deletedFields.length; i++) {
-    console.log(deletedFields[i]);
-    let f: any = deletedFields[i];
-    if (f.id) {
-      await axios
-        .put(`/api/forms/field/${f.id}`, {
-          archive: is_archive,
+    await axios
+        .post("/api/forms", {
+            team_id: idTeam,
         })
         .catch((err) => {
-          if (err.response) {
-            responseMsg.value = err.response.data.message;
-          }
+            if (err.response) {
+                responseMsg.value = err.response.data.message;
+            }
         });
-    }
-  }
 }
 
-//  async function updateForm(id: number, fields_id: string) {
+async function archiveFields(deletedFields: IFormField[], is_archive: boolean) {
+    for (const element of deletedFields) {
+        let f = element;
+        if (f.id) {
+            await axios
+                .put(`/api/forms/field/${f.id}`, {
+                    archive: is_archive,
+                })
+                .catch((err) => {
+                    if (err.response) {
+                        responseMsg.value = err.response.data.message;
+                    }
+                });
+        }
+    }
+}
 
-// responseMsg.value = "сохранено";
-// await axios.put("/api/forms/" + id, {
-//     fields_id: fields_id,
-// })
-//     .catch((err) => {
-//         if (err.response) {
-//             responseMsg.value = err.response.data.message
-//         }
-//     })
-// }
 </script>
 
 <style lang="scss" scoped>
@@ -310,7 +290,7 @@ async function archiveFields(deletedFields: [], is_archive: boolean) {
     height: 80px;
 
     color: #ffffff;
-    background:var(--second-accept);
+    background: var(--second-accept);
   }
 }
 </style>
